requirements.txt
flask
python-dotenv
requests
PyQt5
matplotlib